pyproject.toml
[tool.poetry]
name = "drive-peripheral"
version = "0.1.0"
description = "BLE Peripheral to drive remote control car"
authors = ["Shinya Ishida <shinyaishida@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "drive_peripheral"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pybleno = "^0.11"
python-dotenv = "^0.21.0"

[tool.poetry.group.test]

[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
